fixes issue of swapped activity and switch routes/controllers, tested and all is good now
